// generated/src/app/api-models/subscriber-snapshot-arc.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SubscriberSnapshotArc {
  seqSubsId: number;
  seqSubsUpdate: number;
  changeDateTime: Date;
  effectiveDate: Date;
  termDate: Date;
  seqGroupId: number;
  planCode: string;
  otherStatusFlag: string;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
  familySize: number;
  spouseFlag: string;
  subscriberDob: Date;
  subscriberGender: string;
  riderCode1: string;
  riderCode1NoMbrs: number;
  riderCode2: string;
  riderCode2NoMbrs: number;
  riderCode3: string;
  riderCode3NoMbrs: number;
  riderCode4: string;
  riderCode4NoMbrs: number;
  riderCode5: string;
  riderCode5NoMbrs: number;
  riderCode6: string;
  riderCode6NoMbrs: number;
  riderCode7: string;
  riderCode7NoMbrs: number;
  riderCode8: string;
  riderCode8NoMbrs: number;
  riderCode9: string;
  riderCode9NoMbrs: number;
  riderCode10: string;
  riderCode10NoMbrs: number;
  riderCode11: string;
  riderCode11NoMbrs: number;
  riderCode12: string;
  riderCode12NoMbrs: number;
  riderCode13: string;
  riderCode13NoMbrs: number;
  riderCode14: string;
  riderCode14NoMbrs: number;
  riderCode15: string;
  riderCode15NoMbrs: number;
  riderCode16: string;
  riderCode16NoMbrs: number;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  salary: number;
  medicareStatus: string;
}